requirements.txt
Flask
google-genai
pandas
ortools
python-dotenv